// src-electron/common/types.ts
export interface VoiceStateUser {
	username: string;
	speaking: boolean;
}

export interface VoiceStateInfo {
	channelName: string;
	guildName?: string;
	selfMute: boolean;
	selfDeaf: boolean;
	users?: Array<VoiceStateUser>;
}